Moved CalendarContent.tsx to components folder
